Music: Repaired artists filter in front and aded artists tab
