+++ client/src/pages/client/music/Music.vue
@@ -24,51 +24,9 @@
       <q-tab-panel name="tracks" class="q-pa-none">
         <TracksTab />
       </q-tab-panel>
+
       <q-tab-panel name="artists" class="q-pa-none">
-        <div class="text-h4 q-mb-md">Artists</div>
-        <div class="flex justify-between items-end q-mb-lg">
-          <artists-filter />
-          <q-btn
-            @click="toggleCardMode"
-            color="primary"
-            :icon="cardMode === 'card' ? 'toc' : 'view_cozy'"
-            size="lg"
-            flat
-            round
-          />
-        </div>
-
-        <div v-if="cardMode === 'card'" class="artists-list row items-start q-gutter-md q-mb-lg">
-          <artist-card
-            v-for="artist in artists"
-            :key="artist.id"
-            :artist="artist"
-          />
-          <q-inner-loading :showing="artistsLoading">
-            <q-spinner-gears size="50px" color="primary" />
-          </q-inner-loading>
-        </div>
-
-        <div v-if="cardMode === 'horizontal'" class="row q-gutter-md q-mb-lg">
-          <artist-card-horizontal
-            v-for="artist in artists"
-            :key="artist.id"
-            :artist="artist"
-          />
-          <q-inner-loading :showing="artistsLoading">
-            <q-spinner-gears size="50px" color="primary" />
-          </q-inner-loading>
-        </div>
-        <div class="show-more-button flex justify-center">
-          <q-btn
-            v-if="pagination.hasPages"
-            color="primary"
-            label="Show more"
-            @click="loadMoreArtists"
-            :loading="paginationLoading"
-          >
-          </q-btn>
-        </div>
+        <ArtistsTab />
       </q-tab-panel>
 
       <q-tab-panel name="playlists" class="q-pa-none">
@@ -99,85 +57,34 @@
   </q-page>
 </template>
 <script>
-import ArtistsFilter from "src/components/client/music/ArtistsFilter.vue"
-import ArtistCard from 'src/components/client/music/ArtistCard.vue'
-import ArtistCardHorizontal from 'src/components/client/music/ArtistCardHorizontal.vue'
 import TracksTab from 'src/components/client/music/tabs/TracksTab.vue'
+import ArtistsTab from 'src/components/client/music/tabs/ArtistsTab.vue'
 
 import { ref, onMounted } from "vue";
 import API from "src/utils/api";
 
 export default {
-  components: { ArtistsFilter, ArtistCard, ArtistCardHorizontal, TracksTab },
+  components: { TracksTab, ArtistsTab },
   setup() {
     const tags = ref([])
     let tagsLoading = ref(true)
-    const artists = ref([])
-    let artistsLoading = ref(true)
-    let pagination = ref({
-      perPage: 0,
-      hasPages: false,
-      nextPageUrl: '',
-      prevPageUrl: ''
-    })
-    let paginationLoading = ref(false)
-    let cardMode = ref('horizontal')
 
     const getTags = async () => {
-      const {data} = await API.post('music/tags/tree')
-      tags.value = data.tags
-      tagsLoading.value = false
-    }
-
-    const getArtists = async () => {
-      const {data} = await API.post('music/artists/get')
-
-      artists.value = data.artists
-      pagination.value = data.pagination
-      artistsLoading.value = false
-    }
-
-    const loadMoreArtists = async () => {
-      if (pagination.value.hasPages) {
-        paginationLoading.value = true
-        let obUrl = new URL(pagination.value.nextPageUrl)
-        let cursor = obUrl.searchParams.get("cursor")
-
-        await API.post('music/artists/get', {'cursor': cursor})
-          .then(response => {
-            pagination.value = response.data.pagination
-            artists.value.push(...response.data.artists)
-            paginationLoading.value = false
-          })
-      }
-    }
-
-    const toggleCardMode = () => {
-      if (cardMode.value === 'card') {
-        cardMode.value = 'horizontal'
-      } else {
-        cardMode.value = 'card'
-      }
+      await API.post('music/tags/tree').then(response => {
+        tags.value = response.data.tags
+        tagsLoading.value = false
+      })
     }
 
     onMounted(() => {
       getTags()
-      getArtists()
     })
 
     return {
       tab: ref('artists'),
       tags,
       tagsLoading,
-      artists,
-      pagination,
-      paginationLoading,
-      artistsLoading,
-      cardMode,
-      getTags,
-      getArtists,
-      loadMoreArtists,
-      toggleCardMode
+      getTags
     }
   }
 }
